% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tot_edges}
\alias{tot_edges}
\title{Edges of alliance graph from "Tao Te Programming"}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 124 rows and 3 columns.}
\usage{
tot_edges
}
\description{
\code{tot_edges} is a \link[tibble:tibble]{tibble} with information about chapter
connections in
\href{http://www.burns-stat.com/documents/books/tao-te-programming/}{Tao Te Programming}
book (considered as edges in "alliance graph"). It has three columns:
\itemize{
\item \strong{chapter1} <int> : Number of first chapter in connection.
\item \strong{chapter2} <int> : Number of second chapter in connection.
\item \strong{type} <chr> : Connection type ("ally" for "friendly" connection and
"opponent" for "not friendly" connection).
}
}
\keyword{datasets}
