% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ttp_chapters}
\alias{ttp_chapters}
\title{Chapters of "Tao Te Programming"}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81 rows and 2 columns.}
\usage{
ttp_chapters
}
\description{
\code{ttp_chapters} is a \link[tibble:tibble]{tibble} with information about chapters
of \href{http://www.burns-stat.com/documents/books/tao-te-programming/}{Tao Te Programming}
book. It has two columns:
\itemize{
\item \strong{chapter} <int> : Chapter number.
\item \strong{name} <chr> : Chapter name.
}
}
\keyword{datasets}
